code/requirements.txt
streamlit
openai
azure-cosmos
pandas
numpy
scikit-learn
